services/trust/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.35.3
psycopg2-binary==2.9.9
aio-pika==9.4.1
python-dotenv==1.0.1
